pyproject.toml
[project]
name = "lazydog-office"
version = "0.1.0"
description = "Lazydog.ai Virtual Office - Multi-Agent Social Media Management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "anthropic>=0.28.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "apify-client>=1.7.0",
    "sse-starlette>=2.1.0",
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
